requirements.txt
# Core requirements
pylint>=2.8.0
coverage>=5.5
pytest>=6.2.5
PyYAML>=6.0

# API requirements (optional)
flask>=2.0.0
flask-cors>=3.0.10

# Dashboard requirements (optional)
# flask>=2.0.0
# jinja2>=3.0.0

# Development requirements (optional)
# pytest-cov>=2.12.1
# black>=21.5b2
# isort>=5.9.1
# mypy>=0.812
